Separate hashing into step step data_hash -> hash.
Previously when serializing ProofUnions the type of object was not
hashed if the object was pruned; what was serialized was:

    <type #><pruned?><hash>

Thus it was possible to change the type # without changing the hash of
the deserialized object. By changing the hashing algorithm to:

    H(tag | H(data))

and the serialization to:

    <type #><pruned?><data_hash>

we prevent such modification as changing the type # changes the object
hash.